+++ tests/test_settings_openapi_spec.py
@@ -57,7 +57,7 @@ def test_local_spec_path(app, cli_runner, tmp_path):
     result = cli_runner.invoke(spec_command)
     assert 'openapi' in result.output
     with open(local_spec_path) as f:
-        assert json.loads(f.read()) == app._spec
+        assert json.loads(f.read()) == app.spec
 
 
 @pytest.mark.parametrize('indent', [0, 2, 4])